requirements.txt
bson==0.5.*
fastapi==0.88.*
google-cloud-pubsub==2.13.*
google-cloud-storage==2.7.*
Pillow==9.3.*
python-dotenv==0.21.*
python-multipart==0.0.*
requests==2.28.*
uvicorn==0.20.*
